console compiles again

+++ java_console/models/src/com/rusefi/BinarySearch.java
@@ -1,7 +1,7 @@
 package com.rusefi;
 
 public class BinarySearch {
-    public static int binarySearch(double target, float[] angles) {
+    public static int binarySearch(double target, double[] angles) {
         System.out.println("Testing " + target);
 
         int left = 0;
